Fragment javascript into packaged files
