+++ index.html
@@ -5,7 +5,7 @@
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <link rel="stylesheet" href="index.css">
-        <script src="js/index.js" async defer></script>
+        <script src="js/index.js" type="module" async defer></script>
         <script src="js/form-validation.js" async defer></script>
         <script src="js/form-data.js" async defer></script>
         <title>Daniel's Portfolio</title>
