+++ js/cards.js
@@ -1,0 +1,64 @@
+const cardsArr = [
+  {
+    name: 'card1',
+    image: 'Icons/img-placeholder2.png',
+    title: 'Professional Art Printing Data',
+    text: 'A daily selection of privately personalized reads; no accounts or sign-ups required; has been the industry\'s standard',
+    buttons: {
+      langs: ['HTML', 'Bootstrap', 'Ruby on Rails'],
+      other: [['See Live', '', 'Icons/see-project-icon.svg'], ['See Source', '', 'Icons/source-icon.svg']],
+    },
+  },
+  {
+    name: 'card2',
+    image: 'Icons/img-placeholder2.1.png',
+    title: 'Object-Oriented Programming',
+    text: 'A daily selection of privately personalized reads; no accounts or sign-ups required; has been the industry\'s standard',
+    buttons: {
+      langs: ['HTML', 'Bootstrap', 'Ruby on Rails'],
+      other: [['See Live', '', 'Icons/see-project-icon.svg'], ['See Source', '', 'Icons/source-icon.svg']],
+    },
+  },
+  {
+    name: 'card3',
+    image: 'Icons/img-placeholder1.png',
+    title: 'Obsess No More With Your HTML',
+    text: 'A daily selection of privately personalized reads; no accounts or sign-ups required; has been the industry\'s standard',
+    buttons: {
+      langs: ['HTML', 'Bootstrap', 'Ruby on Rails'],
+      other: [['See Live', '', 'Icons/see-project-icon.svg'], ['See Source', '', 'Icons/source-icon.svg']],
+    },
+  },
+  {
+    name: 'card4',
+    image: 'Icons/img-placeholder2.1.png',
+    title: 'Ruby Just Got Crazy',
+    text: 'A daily selection of privately personalized reads; no accounts or sign-ups required; has been the industry\'s standard',
+    buttons: {
+      langs: ['HTML', 'Bootstrap', 'Ruby on Rails'],
+      other: [['See Live', '', 'Icons/see-project-icon.svg'], ['See Source', '', 'Icons/source-icon.svg']],
+    },
+  },
+  {
+    name: 'card5',
+    image: 'Icons/img-placeholder2.1.png',
+    title: 'Functional Programming? YES',
+    text: 'A daily selection of privately personalized reads; no accounts or sign-ups required; has been the industry\'s standard',
+    buttons: {
+      langs: ['HTML', 'Bootstrap', 'Ruby on Rails'],
+      other: [['See Live', '', 'Icons/see-project-icon.svg'], ['See Source', '', 'Icons/source-icon.svg']],
+    },
+  },
+  {
+    name: 'card6',
+    image: 'Icons/img-placeholder2.1.png',
+    title: 'The Rabbit-Hole To End All Rabbit-Holes Just Got Deep!',
+    text: 'A daily selection of privately personalized reads; no accounts or sign-ups required; has been the industry\'s standard',
+    buttons: {
+      langs: ['HTML', 'Bootstrap', 'Ruby on Rails'],
+      other: [['See Live', '', 'Icons/see-project-icon.svg'], ['See Source', '', 'Icons/source-icon.svg']],
+    },
+  },
+];
+
+export default cardsArr;
